fikset små feil i koden

+++ klasser2.py
@@ -6,9 +6,6 @@ from random import randint
 import math as math
      
 
-
-
-
 class SpillObjekt:
     def __init__(self, x:int, y:int):
         self.x = x
@@ -162,7 +159,7 @@ pg.init()
 font = pg.font.SysFont("Tahoma", 24)
 
 class Knapp:
-  def __init__(self, xPosisjon:int, yPosisjon:int, tekst:str):
+  def __init__(self, xPosisjon:float, yPosisjon:float, tekst:str):
     self.xPosisjon = xPosisjon
     self.yPosisjon = yPosisjon
     self.bredde = len(tekst) * 20
